task a/requirements_task_a.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0 